change daily_tasks.py

+++ daily_tasks.py
@@ -3,8 +3,9 @@ import os
 import secrets
 import shutil
 
-# Backup database
-shutil.copy2('database.db', 'database.db.bak')
+# Backup database if exists
+if os.path.exists('database.db'):
+    shutil.copy2('database.db', 'database.db.bak')
 
 # Generate new secret key
 secret = secrets.token_hex(48)  # 384 bits
@@ -12,6 +13,7 @@ with open('secret_key.txt', 'w') as file:
     file.write(secret)
 
 # rotate logs
-timestamp = datetime.date.strftime(datetime.datetime.now(), "%d-%m-%Y")
-shutil.copy2("logs/application.log", f"logs/{timestamp}-application.log")
-os.remove("logs/application.log")
+if os.path.exists('logs/application.log'):
+    timestamp = datetime.date.strftime(datetime.datetime.now(), "%d-%m-%Y")
+    shutil.copy2("logs/application.log", f"logs/{timestamp}-application.log")
+    os.remove("logs/application.log")
